+++ indico/web/rh_test.py
@@ -9,6 +9,7 @@ from contextlib import nullcontext as does_not_raise
 from unittest.mock import MagicMock
 
 import pytest
+from flask import request
 from werkzeug.exceptions import Forbidden
 
 from indico.core import signals
@@ -128,3 +129,54 @@ def test_cors_headers_set_after_error(app):
         resp = client.get('/test/cors/error', headers=headers)
         # Ensure the CORS headers are set even if an error was raised
         assert resp.headers['Access-Control-Allow-Origin'] == 'localhost'
+
+
+def test_normalize_simple(app, test_client):
+    @make_view_func
+    class DummyRH(RH):
+        normalize_url_spec = {'locators': {lambda self: {'id': 42}}}
+
+        def _process(self):
+            assert request.view_args['id'] == 42
+            return str(request.view_args['id'])
+
+    app.add_url_rule('/test/normalize/simple/<int:id>', 'test_normalize_simple', DummyRH)
+
+    resp = test_client.get('/test/normalize/simple/1', follow_redirects=True)
+    assert resp.request.path == '/test/normalize/simple/42'
+
+
+class _LateDummyRH(RH):
+    normalize_url_spec = {'locators': {lambda self: {'id': 42}}, 'preserved_args': {'slug'}}
+    normalize_url_spec_late = {'locators': {lambda self: {'id': 42, 'slug': 'meow'}}}
+
+    def _check_access(self):
+        if not self.authorized:
+            raise Forbidden
+
+    def _process(self):
+        assert request.view_args['id'] == 42
+        return str(request.view_args['id'])
+
+
+@pytest.mark.parametrize('authorized', (False, True))
+@pytest.mark.parametrize('suffix', ('', '-meow', '-foo'))
+def test_normalize_late(app, test_client, dummy_user, authorized, suffix):
+    _LateDummyRH.authorized = authorized
+    app.add_url_rule('/test/normalize/late/<int:id>', 'test_normalize_late', make_view_func(_LateDummyRH))
+    app.add_url_rule('/test/normalize/late/<int:id>-<slug>', 'test_normalize_late', make_view_func(_LateDummyRH))
+
+    with test_client.session_transaction() as sess:
+        sess.set_session_user(dummy_user)  # avoid redirect to login page
+    resp = test_client.get(f'/test/normalize/late/1{suffix}', follow_redirects=True)
+    history = [x.location for x in resp.history]
+
+    if authorized:
+        assert resp.request.path == '/test/normalize/late/42-meow'
+        # Using a set here because the order is not really significant, and it's easiest way to deduplicate.
+        assert set(history) == {f'/test/normalize/late/42{suffix}', '/test/normalize/late/42-meow'}
+    else:
+        assert resp.request.path == f'/test/normalize/late/42{suffix}'
+        assert history == [f'/test/normalize/late/42{suffix}']
+
+    # TODO: Add more tests for other URL normalization functionality